Rename test file.
