+++ hugr-passes/src/const_fold/test.rs
